% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bry_boschan.R
\name{bry_boschan}
\alias{bry_boschan}
\title{This function determines the phase based on the Bry-Boschan Algorithm}
\usage{
bry_boschan(dat, country, mincycle = 5, minphase = 2)
}
\description{
This function determines the phase based on the Bry-Boschan Algorithm
}
