% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS.R
\name{det_phase}
\alias{det_phase}
\title{This function determines whether a period is boom, bust, or neither based on deviation}
\usage{
det_phase(dat, lambda)
}
\description{
This function determines whether a period is boom, bust, or neither based on deviation
}
