% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS.R
\name{enforce_minphase}
\alias{enforce_minphase}
\title{This function makes sure the minimum phase length is met.}
\usage{
enforce_minphase(x, minphase, gap)
}
\description{
This function makes sure the minimum phase length is met.
}
