% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersect.R
\name{find_intersect}
\alias{find_intersect}
\title{This function finds the intersection by omitting rows containing NAs}
\usage{
find_intersect(x, country)
}
\arguments{
\item{x}{dataframe containing information on 2/3 data series}

\item{country}{the country we'd like to look at}
}
\description{
This function finds the intersection by omitting rows containing NAs
}
