% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersect.R
\name{find_start_end}
\alias{find_start_end}
\title{This function finds the start and end date of a country given a data series}
\usage{
find_start_end(country, dat, dat_name)
}
\arguments{
\item{country}{country name}

\item{dat}{dataset}

\item{dat_name}{name of the series}
}
\description{
This function finds the start and end date of a country given a data series
}
