% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersect.R
\name{find_start_end_all}
\alias{find_start_end_all}
\title{This function is a helper that calls find_start_end}
\usage{
find_start_end_all(country_list, dat, dat_name)
}
\arguments{
\item{country_list}{list of country names}

\item{dat}{dataset}

\item{dat_name}{name of the series}
}
\description{
This function is a helper that calls find_start_end
}
