% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dat.R
\name{merge_dat}
\alias{merge_dat}
\title{This function merges the read data.
It can either merge two with the given keys, or merge all three}
\usage{
merge_dat(x, first = "mp", second = "pp", all = FALSE)
}
\arguments{
\item{x}{a list of dataframes containing the read data}

\item{first}{first dataframe to merge}

\item{second}{second dataframe to merge}

\item{all}{boolean indicating whether all three are merged}
}
\description{
This function merges the read data.
It can either merge two with the given keys, or merge all three
}
