% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersect.R
\name{mp_reduce_range}
\alias{mp_reduce_range}
\title{This function turn the entries into NA if not covered in coverage years}
\usage{
mp_reduce_range(cy, mp)
}
\arguments{
\item{mp}{macroprudential action data}

\item{x}{coverage years}
}
\description{
This function turn the entries into NA if not covered in coverage years
}
