% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_cy}
\alias{read_cy}
\title{This function reads in coverage years for macroprudential actions.
For end dates that are not the last day of the quarter, we set them
to the quarters before.}
\usage{
read_cy(path)
}
\arguments{
\item{path}{is the file path}
}
\description{
This function reads in coverage years for macroprudential actions.
For end dates that are not the last day of the quarter, we set them
to the quarters before.
}
