% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_mp}
\alias{read_mp}
\title{This function reads in the macroprudential policy actions and cleans
the date formatting}
\usage{
read_mp(x, path)
}
\arguments{
\item{x}{is the specific prudential policy}

\item{path}{is the file path}
}
\description{
This function reads in the macroprudential policy actions and cleans
the date formatting
}
