% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersect.R
\name{reduce_coverage}
\alias{reduce_coverage}
\title{This function turn the entries into NA if not covered in coverage years}
\usage{
reduce_coverage(dates, dat, coverage_name = "all")
}
\arguments{
\item{dates}{coverage years}

\item{dat}{dataframe being reduced}

\item{coverage_name}{which coverage to look at. We are usually using 'all'}

\item{country}{which country to look at}
}
\description{
This function turn the entries into NA if not covered in coverage years
}
