% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{select_pp}
\alias{select_pp}
\title{This function selects the appropriate price data for the given country}
\usage{
select_pp(country, pp)
}
\arguments{
\item{country}{is the country name}

\item{pp}{is the pp data returned by `read_pp`}
}
\description{
This function selects the appropriate price data for the given country
}
