% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS.R
\name{smooth_ks}
\alias{smooth_ks}
\title{This function calculated the growth rate and take the 3 year moving average}
\usage{
smooth_ks(dat, country)
}
\description{
This function calculated the growth rate and take the 3 year moving average
}
