% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersect.R
\name{summarize_mp}
\alias{summarize_mp}
\title{This function summarizes the number of mp actions for a given country}
\usage{
summarize_mp(mp_hc, country)
}
\arguments{
\item{mp_hc}{the dataset merging macroprudential data and housing credit data}

\item{country}{the country we are examining}
}
\description{
This function summarizes the number of mp actions for a given country
}
