% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{take_log}
\alias{take_log}
\title{Takes the log of all columns but date}
\usage{
take_log(x)
}
\arguments{
\item{x}{the data frame we want to take the log of}
}
\description{
Takes the log of all columns but date
}
