% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_hc}
\alias{read_hc}
\title{This function reads in the housing market data and cleans
the date formatting}
\usage{
read_hc(x, cpi)
}
\arguments{
\item{x}{is the filepath to the credit data}

\item{cpi}{is the cpi data we deflate with}
}
\description{
This function reads in the housing market data and cleans
the date formatting
}
