% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_pp}
\alias{read_pp}
\title{This function reads in the property price data and cleans
the date formatting}
\usage{
read_pp(pps, ppl, pp_saar, cpi, country_list)
}
\arguments{
\item{cpi}{is the cpi data}

\item{country_list}{is the list of countries}

\item{x}{is the filepath to the price data}
}
\description{
The returned data frame is deflated and in log form.
}
